手机验证码

+++ app/Http/Controllers/Index/ReplyController.php
@@ -18,6 +18,7 @@ class ReplyController extends Controller
     }
 
     public function destroy(Reply $reply){
+        $this->authorize('handle', $reply);
         $reply->delete();
         return back()->with('success', '回复成功删除');
     }
